UI: Monitoring charts added timewindow per each chart; use precise Chart.js types; add monitoring-chart.component.
