+++ ui-ngx/src/app/modules/home/pages/monitoring/monitoring-chart.component.html
@@ -1,0 +1,143 @@
+<!--
+
+    Copyright © 2016-2025 The Thingsboard Authors
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+-->
+<div class="flex flex-1 flex-col gap-4">
+    <div tb-fullscreen
+      [fullscreen]="isFullscreen"
+      [fullscreenBackgroundStyle]="{'background-color':'#fff'}">
+      <div [id]="chartType() + 'container'" #chartElement class="mat-padding chart-container">
+        <mat-divider [class.!hidden]="!isFullscreen"></mat-divider>
+        <section class="flex flex-row items-center justify-between" [style]="{'padding-right': isFullscreen ? '20px' : 0}">
+          <tb-timewindow [(ngModel)]="timewindow"
+                         (ngModelChange)="onTimewindowChange()"
+                         asButton
+                         flatButton
+                         tooltipPosition="left"
+                         aggregation="true">
+          </tb-timewindow>
+          <span class="flex flex-1"></span>
+          <span [style]="{paddingLeft: chartType() === 'processedBytes' ? '170px' : 0}">{{ chartTypeTranslationMap.get(chartType()) | translate }}</span>
+          <span class="flex flex-1"></span>
+          @if (chartType() === 'processedBytes') {
+            <tb-toggle-header (valueChange)="processedBytesUnitTypeChanged($event)"
+              [value]="currentDataSizeUnitType"
+              style="padding-right: 12px"
+              [name]="chartType()"
+              useSelectOnMdLg="true"
+              appearance="fill">
+              @for (type of dataSizeUnitType; track type) {
+                <tb-toggle-option [value]="type">
+                  {{ dataSizeUnitTypeTranslationMap.get(type) }}
+                </tb-toggle-option>
+              }
+            </tb-toggle-header>
+          }
+          <mat-icon [matTooltip]="chartTooltip(chartType())" class="material-icons" style="cursor: default">
+            info_outline
+          </mat-icon>
+          <button type="button"
+            mat-icon-button class="tb-fullscreen-button tb-mat-40"
+            (click)="onFullScreen(chartType())"
+            matTooltip="{{(isFullscreen ? 'fullscreen.exit' : 'fullscreen.expand') | translate}}"
+            matTooltipPosition="above">
+            <mat-icon>{{ isFullscreen ? 'fullscreen_exit' : 'fullscreen' }}</mat-icon>
+          </button>
+        </section>
+        <section [style]="{'height': chartContainerHeight}">
+          <canvas [id]="chartType() + chartPage" [height]="chartHeight"></canvas>
+        </section>
+        @if (showLegend) {
+          <section class="tb-time-series-chart-legend">
+            <ng-container *ngTemplateOutlet="legend"></ng-container>
+          </section>
+        }
+        <mat-divider [class.!hidden]="!isFullscreen"></mat-divider>
+      </div>
+    </div>
+</div>
+
+<ng-template #legend>
+  <table class="tb-time-series-chart-legend-table vertical">
+    <thead>
+      <tr>
+        <th></th>
+        @if (legendConfig.showMin === true) {
+          <th class="tb-time-series-chart-legend-type-label right">{{ 'legend.Min' | translate }}</th>
+        }
+        @if (legendConfig.showMax === true) {
+          <th class="tb-time-series-chart-legend-type-label right">{{ 'legend.Max' | translate }}</th>
+        }
+        @if (legendConfig.showAvg === true) {
+          <th class="tb-time-series-chart-legend-type-label right">{{ 'legend.Avg' | translate }}</th>
+        }
+        @if (legendConfig.showTotal === true && !totalOnly()) {
+          <th class="tb-time-series-chart-legend-type-label right">{{ 'legend.Total' | translate }}</th>
+        }
+        @if (legendConfig.showLatest === true) {
+          <th class="tb-time-series-chart-legend-type-label right">{{ 'legend.Latest' | translate }}</th>
+        }
+      </tr>
+    </thead>
+    <tbody>
+      @for (legendKey of legendKeys; track legendKey; let index = $index) {
+        <tr>
+          <th>
+            <ng-container *ngTemplateOutlet="legendItem; context:{legendKey: legendKey, left: true}"></ng-container>
+          </th>
+          @if (legendConfig.showMin === true && !legendKey.dataKey.hidden) {
+            <td class="tb-time-series-chart-legend-value"
+              [innerHTML]="legendValue(index, 'min') | safe: 'html'">
+            </td>
+          }
+          @if (legendConfig.showMax === true && !legendKey.dataKey.hidden) {
+            <td class="tb-time-series-chart-legend-value"
+              [innerHTML]="legendValue(index, 'max') | safe: 'html'">
+            </td>
+          }
+          @if (legendConfig.showAvg === true && !legendKey.dataKey.hidden) {
+            <td class="tb-time-series-chart-legend-value"
+              [innerHTML]="legendValue(index, 'avg') | safe: 'html'">
+            </td>
+          }
+          @if (legendConfig.showTotal === true && !legendKey.dataKey.hidden && !totalOnly()) {
+            <td
+              class="tb-time-series-chart-legend-value"
+              [innerHTML]="legendValue(index, 'total') | safe: 'html'">
+            </td>
+          }
+          @if (legendConfig.showLatest === true && !legendKey.dataKey.hidden) {
+            <td class="tb-time-series-chart-legend-value"
+              [innerHTML]="legendValue(index, 'latest') | safe: 'html'">
+            </td>
+          }
+        </tr>
+      }
+    </tbody>
+  </table>
+</ng-template>
+<ng-template #legendItem let-legendKey="legendKey" let-left="left">
+  <div class="tb-time-series-chart-legend-item"
+    [class.left]="left">
+    <div class="tb-time-series-chart-legend-item-label"
+      (mouseenter)="onLegendKeyEnter(legendKey)"
+      (mouseleave)="onLegendKeyLeave(legendKey)"
+      (click)="toggleLegendKey(legendKey)">
+      <div class="tb-time-series-chart-legend-item-label-circle" [style]="{background: !legendKey.dataKey.hidden ? legendKey.dataKey.color : null}"></div>
+      <div class="tb-time-series-chart-legend-item-label-text" [innerHTML]="legendKey.dataKey.label | safe: 'html'"></div>
+    </div>
+  </div>
+</ng-template>
